Refine: Topics and TweetFeed refactor
