+++ src/mocks/topics.json
@@ -1,0 +1,17 @@
+{
+    "name": "Qué está pasando",
+    "items": [
+      {
+        "topic": "Tecnología · Tendencia", "name": "GitHub Copilot", "tweets": "5.436", "image": "https://github.blog/wp-content/uploads/2022/06/Copilot.jpeg?resize=1200%2C630"
+      },
+      {
+        "topic": "Tendencia en México", "name": "TikTok", "tweets": "539 mil"
+      },
+      {
+        "topic": "Tendencia en México", "name": "#DoctorStrange", "tweets": "13,8 mil"
+      },
+      {
+        "topic": "Entretenimiento · Tendencia", "name": "One Piece", "tweets": "112 mil", "image": "https://lafrikileria.com/blog/wp-content/uploads/2021/07/one-piece-logo.jpg"
+      }
+    ]
+  }
